Remove rss helper
